+++ game/missiongenerator/aircraft/waypoints/refuel.py
@@ -1,16 +1,12 @@
 from dcs.point import MovingPoint
 from dcs.task import RefuelingTaskAction, ControlledTask
 
-from game.theater import OffMapSpawn
 from .pydcswaypointbuilder import PydcsWaypointBuilder
 
 
 class RefuelPointBuilder(PydcsWaypointBuilder):
     def add_tasks(self, waypoint: MovingPoint) -> None:
-        offmap = isinstance(self.flight.arrival, OffMapSpawn)
-        if self.flight.divert:
-            offmap |= isinstance(self.flight.divert, OffMapSpawn)
-        if not offmap:
+        if not self.ai_despawn(waypoint, True):
             refuel = ControlledTask(RefuelingTaskAction())
             refuel.start_probability(10)
             waypoint.add_task(refuel)